added explicit cast since a ambiguity arose with gcc3.2
git-svn-id: svn://svn.code.sf.net/p/libvideogfx/code/trunk@78 89ed393f-2632-0410-9d69-91a09fb2310b

+++ libvideogfx/libvideogfx/graphics/draw/draw.hh
@@ -496,7 +496,7 @@ namespace videogfx {
     dxp = x1-x0;
     dyp = y1-y0;
 
-    norm = ::sqrt(dxp*dxp+dyp*dyp);
+    norm = ::sqrt(((double)(dxp*dxp+dyp*dyp));
 
     if (norm>0.0)
       {
